+++ src/ovc_enc/spiht/skip/encoder.cpp
@@ -20,7 +20,7 @@ void skip_spiht_encoder::encode(matrix<double> in_matrix, ovc_spiht_config in_co
 	}
 }
 
-void skip_spiht_encoder::flush(uint8_t** out_bits, int* out_step)
+void skip_spiht_encoder::flush(uint8_t** out_bits, uint16_t* out_step)
 {
 	*out_bits = new uint8_t[bitstream->occupancy()];
 	memcpy(*out_bits, bitstream->data(), bitstream->occupancy());
