+++ src/ovc_enc/spiht/skip/encoder.h
@@ -6,5 +6,5 @@ class skip_spiht_encoder : public ovc_spiht_encoder
 {
 public:
 	virtual void encode(matrix<double> in_matrix, ovc_spiht_config in_config) override;
-	virtual void flush(uint8_t** out_bits, int* out_step) override;
+	virtual void flush(uint8_t** out_bits, uint16_t* out_step) override;
 };
